refactor and fix circular dep when importing base class from .
