+++ code/src/nuvla/api/resources/callback.py
@@ -1,4 +1,5 @@
-from . import ResourceBase
+
+from .base import ResourceBase
 
 
 class Callback(ResourceBase):
